worker/requirements.txt
pika
opencv-python-headless
psycopg2-binary
boto3
